+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Installer script that can be used via `curl -s "https://get.kamu.dev" | sh`
+
 ## [0.103.0] - 2022-12-23
 ### Changed
 - Major upgrade to latest versions of `flatbuffers`, `arrow`, `datafusion`, and many more dependencies.